Fix github action to pull enough information
In the original version of this, the `checkout` action does the *bare*
minimum to test things. As such, `git rev-list` was failing because it
didn't understand the reference to `main`. This updates how and what we
checkout based off of the horribly undocumented features of GitHub
actions. This will check all commits in a PR's branch rather than just
the latest commit on a PR.